+++ src/ProgressBar/index.ts
@@ -1,0 +1,4 @@
+import ProgressBar from './progress_bar';
+
+
+export default ProgressBar;